requirements.txt
line-bot-sdk==3.16.3
Flask==2.0.1
Werkzeug==2.0.3
python-dotenv==0.19.0
google-generativeai==0.3.2
Pillow==10.2.0
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0 